+++ config.py
@@ -1,0 +1,4 @@
+# config the site
+class Config:
+    static_dir = "/static"
+    base_url = ""
