+++ static/css/main.css
@@ -1,1 +1,189 @@
+* {
+	box-sizing: border-box;
+}
+
+body {
+    margin: 0px;
+	background-repeat: repeat;
+	background-position: center top;
+	background-attachment: fixed;
+	background-size: cover;
+}
+
+.content {
+    margin: 0 auto;
+    padding: 60px 10px;
+    max-width: 975px;
+    height: 100%;
+    min-height: 100vh;
+    background: white;
+    opacity: 0.8;
+}
+
+.heading {
+	padding: 10px;
+}
+
+.paragraph {
+	padding: 10px;
+}
+
+.bdg-sect {
+	padding-top: 100px;
+	padding-bottom: 100px;
+	font-family: Helvetica, serif;
+	background-color: #fafafa;
+}
+
+.navbar-items-c {
+	display: inline-block;
+	float: right;
+}
+
+.navbar {
+    position: fixed;
+    width: 100%;
+    z-index: 1;
+}
+
+.navbar-container {
+	margin: 0 auto;
+	width: 100%;
+	max-width: 975px;
+}
+
+.navbar-container::after {
+	content: "";
+	clear: both;
+	display: block;
+}
+
+.navbar-brand {
+	vertical-align: top;
+	display: inline-block;
+	padding: 5px;
+	min-height: 50px;
+	min-width: 50px;
+	color: inherit;
+	text-decoration: none;
+}
+
+.navbar-menu {
+	padding: 10px 0;
+	display: block;
+	float: right;
+	margin: 0;
+}
+
+.navbar-menu-link {
+	margin: 0;
+	color: inherit;
+	text-decoration: none;
+	display: inline-block;
+	padding: 10px 15px;
+}
+
+.navbar-burger {
+	margin: 10px 0;
+	width: 45px;
+	padding: 5px 10px;
+	display: none;
+	float: right;
+	cursor: pointer;
+}
+
+.navbar-burger-line {
+	padding: 1px;
+	background-color: white;
+	margin: 5px 0;
+}
+
+#i6nx {
+	background-image: linear-gradient(rgba(0, 0, 0, 0.47), rgba(0, 0, 0, 0.47)), url(https://cdn.discordapp.com/attachments/701312825670500365/721304506847068250/srb20051.png);
+	background-repeat: repeat, no-repeat;
+	background-position: left top, center center;
+	background-attachment: scroll, scroll;
+	background-size: auto, cover;
+}
+
+#iofq {
+	color: #ffffff;
+}
+
+#circuit_header {
+    font-size: 41px;
+}
+
+#ihbt {
+	color: #ffffff;
+}
+
+#ijh6 {
+	color: #ffffff;
+}
+
+#iwidy {
+	background-image: url(https://cdn.discordapp.com/attachments/701312825670500365/721309227003543572/srb20057.png);
+	background-repeat: repeat;
+	background-position: left top;
+	background-attachment: scroll;
+	background-size: auto;
+}
+
+#i5pmk {
+	color: #ffffff;
+}
+
+.navbar,.navbar-container {
+	background-image: linear-gradient(#3d3d3d, #3d3d3d);
+	background-repeat: repeat;
+	background-position: left top;
+	background-attachment: scroll;
+	background-size: auto;
+}
+
+#i7s3x {
+	color: #ffffff;
+}
+
+#i64ju {
+	color: #ffffff;
+}
+
+#ihy3g {
+	color: #ffffff;
+}
+
+#i1g6j {
+	color: #ffffff;
+}
+
+#whatiscircuit {
+    color: white;
+   text-shadow: 2px 2px black;
+}
+
+#srb2_link {
+    background: white;
+    padding: 3px;
+    border-radius: 2px;
+    opacity: 0.5;
+}
+
+@media (max-width: 768px) {
+	.navbar-burger {
+		display: block;
+	}
+	.navbar-items-c {
+		display: none;
+		width: 100%;
+	}
+	.navbar-menu {
+		width: 100%;
+	}
+	.navbar-menu-link {
+		display: block;
+	}
+}
+
