+++ static/js/main.js
@@ -1,4 +1,4 @@
-var api_url = window.location.protocol + "//" + window.location.host + '/api'
+var api_url = window.location.protocol + "//" + window.location.host + '/highscores/api'
 
 var form = document.getElementById("highscores_form");
 form.addEventListener("submit", function (event) {
