+++ templates/search.html
@@ -1,91 +1,88 @@
-<html>
-<head>
-    <link href="{{config.static_dir}}/css/main.css" type="stylesheet">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-</head>
-<body>
-    <form id="highscores_form" action="/api">
-        <table>
-            <tr>
-                <td>
-                    <label for="username">Username</label>
-                </td>
-                <td>
-                    <input id="username" list="users" name="username" type="text">
-                    <datalist id="users">
-                        {% for user in users: %}
-                        <option value="{{user}}">{{user}}</option>
-                        {% endfor %}
-                    </datalist>
-                    </input>
-                </td>
-            </tr>
-            <tr>
-                <td>
-                    <label for="map_id">Map</label>
-                </td>
-                <td>
-                    <select id="map_id"  name="map_id"  type="select" onchange="submit_form()">
-                        <option value="">All maps</option>
-                    {% for map in maps: %}
-                        <option value="{{map.id}}">{{map.name}}</option>
+{% extends "highscores.html" %}
+{% block content %}
+<form id="highscores_form" action="/api">
+    <table>
+        <tr>
+            <td>
+                <label for="username">Username</label>
+            </td>
+            <td>
+                <input id="username" list="users" name="username" type="text">
+                <datalist id="users">
+                    {% for user in users: %}
+                    <option value="{{user}}">{{user}}</option>
                     {% endfor %}
-                    </select>
-                </td>
-            </tr>
-            <tr>
-                <td>
-                    <label for="skin">Skin</label>
-                </td>
-                <td>
-                    <select id="skin" name="skin" type="select" onchange="submit_form()">
-                        <option value="">All skins</option>
-                        {% for skin in skins: %}
-                        <option value="{{skin}}">{{skin}}</option>
-                        {% endfor %}
-                    </select>
-                </td>
-            </tr>
-            <tr>
-                <td>
-                    <label for="order">Order by</label>
-                </td>
-                <td>
-                    <select id="order" name="order" type="select" onchange="submit_form()">
-                        {% for column in columns: %}
-                        <option value="{{column}}" {% if column == 'time' %} selected="selected" {% endif %}>{{column}}</option>
-                        {% endfor %}
-                    </select>
-                    <select id="descending" name="descending" type="select" onchange="submit_form()">
-                        <option value="">Ascending</option>
-                        <option value="1">Descending</option>
-                    </select>
-                </td>
-            </tr>
-            <tr>
-                <td>
-                    <label for="limit">#Results</label>
-                </td>
-                <td>
-                    <input id="limit" name="limit" type="number" value="20"></input>
-                </td>
-            </tr>
-        </table>
-        <button type="submit">Search</button>
-    </form>
-    <table border=1>
-        <thead>
-            <tr>
-                <th>Username</th>
-                <th>Map</th>
-                <th>Skin</th>
-                <th>Time</th>
-                <th>Timestamp</th>
-            </tr>
-        </thead>
-        <tbody id="highscores_table">
-        </tbody>
+                </datalist>
+                </input>
+            </td>
+        </tr>
+        <tr>
+            <td>
+                <label for="map_id">Map</label>
+            </td>
+            <td>
+                <select id="map_id"  name="map_id"  type="select" onchange="submit_form()">
+                    <option value="">All maps</option>
+                {% for map in maps: %}
+                    <option value="{{map.id}}">{{map.name}}</option>
+                {% endfor %}
+                </select>
+            </td>
+        </tr>
+        <tr>
+            <td>
+                <label for="skin">Skin</label>
+            </td>
+            <td>
+                <select id="skin" name="skin" type="select" onchange="submit_form()">
+                    <option value="">All skins</option>
+                    {% for skin in skins: %}
+                    <option value="{{skin}}">{{skin}}</option>
+                    {% endfor %}
+                </select>
+            </td>
+        </tr>
+        <tr>
+            <td>
+                <label for="order">Order by</label>
+            </td>
+            <td>
+                <select id="order" name="order" type="select" onchange="submit_form()">
+                    {% for column in columns: %}
+                    <option value="{{column}}" {% if column == 'time' %} selected="selected" {% endif %}>{{column}}</option>
+                    {% endfor %}
+                </select>
+                <select id="descending" name="descending" type="select" onchange="submit_form()">
+                    <option value="">Ascending</option>
+                    <option value="1">Descending</option>
+                </select>
+            </td>
+        </tr>
+        <tr>
+            <td>
+                <label for="limit">#Results</label>
+            </td>
+            <td>
+                <input id="limit" name="limit" type="number" value="20"></input>
+            </td>
+        </tr>
     </table>
-</body>
+    <button type="submit">Search</button>
+</form>
+<table border=1>
+    <thead>
+        <tr>
+            <th>Username</th>
+            <th>Map</th>
+            <th>Skin</th>
+            <th>Time</th>
+            <th>Timestamp</th>
+        </tr>
+    </thead>
+    <tbody id="highscores_table">
+    </tbody>
+</table>
+{% endblock %}
+{% block scripts %}
 <script src="{{config.static_dir}}/js/main.js"></script>
-</html>
+{% endblock %}
